requirements.txt
tvdatafeed
python-dotenv
requests
pandas